refactor: refactoring views
