+++ frontend/lib/views/soon_view.dart
