+++ frontend/lib/widgets/home_header.dart
@@ -1,0 +1,123 @@
+import 'package:flutter/material.dart';
+import 'package:frontend/main.dart';
+import 'package:frontend/models/user.dart';
+import 'package:frontend/services/auth.dart';
+import 'package:frontend/pages/my_offers_page.dart';
+
+class HomeHeader extends StatelessWidget {
+  final String userName;
+  final UserProfile? userData;
+  final bool isDark;
+
+  const HomeHeader({
+    super.key,
+    required this.userName,
+    this.userData,
+    required this.isDark,
+  });
+
+  String _getInitials(String name) {
+    List<String> names = name.trim().split(" ");
+    String initials = "";
+    if (names.isNotEmpty) {
+      if (names[0].isNotEmpty) initials += names[0][0];
+      if (names.length > 1 && names[names.length - 1].isNotEmpty) {
+        initials += names[names.length - 1][0];
+      }
+    }
+    return initials.toUpperCase();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final displayName = userData?.name ?? userName;
+
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text(
+              'Olá,',
+              style: TextStyle(fontSize: 16, color: theme.colorScheme.onSurfaceVariant),
+            ),
+            Text(
+              displayName,
+              style: TextStyle(
+                fontSize: 24,
+                fontWeight: FontWeight.bold,
+                color: theme.colorScheme.onSurface,
+              ),
+            ),
+          ],
+        ),
+        Row(
+          children: [
+            IconButton(
+              icon: Icon(
+                isDark ? Icons.light_mode_outlined : Icons.dark_mode_outlined,
+                color: theme.colorScheme.onSurface,
+              ),
+              onPressed: () {
+                themeNotifier.value = isDark ? ThemeMode.light : ThemeMode.dark;
+              },
+            ),
+            const SizedBox(width: 8),
+            PopupMenuButton<String>(
+              offset: const Offset(0, 56),
+              shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+              onSelected: (value) async {
+                if (value == 'offers') {
+                  Navigator.of(context).push(
+                    MaterialPageRoute(builder: (context) => const MyOffersView()),
+                  );
+                } else if (value == 'logout') {
+                  await AuthService.signOut();
+                  if (context.mounted) {
+                    Navigator.of(context)
+                        .pushNamedAndRemoveUntil('/login', (route) => false);
+                  }
+                }
+              },
+              itemBuilder: (context) => [
+                PopupMenuItem<String>(
+                  value: 'offers',
+                  child: Row(
+                    children: [
+                      Icon(Icons.local_offer_outlined, color: theme.colorScheme.onSurface, size: 20),
+                      const SizedBox(width: 12),
+                      const Text('Minhas Ofertas'),
+                    ],
+                  ),
+                ),
+                PopupMenuItem<String>(
+                  value: 'logout',
+                  child: Row(
+                    children: [
+                      Icon(Icons.logout, color: theme.colorScheme.onSurface, size: 20),
+                      const SizedBox(width: 12),
+                      const Text('Sair'),
+                    ],
+                  ),
+                ),
+              ],
+              child: CircleAvatar(
+                radius: 24,
+                backgroundColor: const Color(0xFF00A84E),
+                child: Text(
+                  _getInitials(displayName),
+                  style: const TextStyle(
+                    color: Colors.white,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+              ),
+            ),
+          ],
+        ),
+      ],
+    );
+  }
+}
