+++ frontend/lib/widgets/market_offer_card.dart
@@ -1,0 +1,184 @@
+import 'package:flutter/material.dart';
+import 'package:frontend/models/offer.dart';
+import 'package:intl/intl.dart';
+
+class MarketOfferCard extends StatelessWidget {
+  final OfferWithId offer;
+  final VoidCallback onBuyPressed;
+
+  const MarketOfferCard({
+    super.key,
+    required this.offer,
+    required this.onBuyPressed,
+  });
+
+  String _formatCurrency(int cents) {
+    final format = NumberFormat.currency(
+      locale: 'pt_BR',
+      symbol: 'R\$',
+    );
+    return format.format(cents / 100);
+  }
+
+  Color _getStatusColor(OfferStatus status) {
+    switch (status) {
+      case OfferStatus.open:
+        return Colors.green;
+      case OfferStatus.accepted:
+        return Colors.blue;
+      case OfferStatus.cancelled:
+        return Colors.red;
+      case OfferStatus.expired:
+        return Colors.orange;
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Container(
+      margin: const EdgeInsets.only(bottom: 16),
+      padding: const EdgeInsets.all(16),
+      decoration: BoxDecoration(
+        color: theme.colorScheme.surface,
+        borderRadius: BorderRadius.circular(16),
+        border: Border.all(color: theme.dividerColor.withOpacity(0.1)),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.04),
+            blurRadius: 10,
+            offset: const Offset(0, 4),
+          ),
+        ],
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      offer.startupName,
+                      style: TextStyle(
+                        fontWeight: FontWeight.bold,
+                        fontSize: 18,
+                        color: theme.colorScheme.onSurface,
+                      ),
+                    ),
+                    const SizedBox(height: 2),
+                    Text(
+                      'Vendedor: ${offer.seller.name}',
+                      style: TextStyle(
+                        fontSize: 12,
+                        color: theme.colorScheme.onSurfaceVariant,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              _buildStatusBadge(offer.status),
+            ],
+          ),
+          const SizedBox(height: 12),
+          Row(
+            children: [
+              _buildInfoItem(
+                context,
+                Icons.token_outlined,
+                '${offer.qtdTokens} tokens',
+              ),
+              const SizedBox(width: 16),
+              _buildInfoItem(
+                context,
+                Icons.monetization_on_outlined,
+                _formatCurrency(offer.tokenPriceCents),
+                label: 'cada',
+              ),
+            ],
+          ),
+          Divider(height: 24, color: theme.dividerColor.withOpacity(0.1)),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    'Total da Oferta',
+                    style: TextStyle(color: theme.colorScheme.onSurfaceVariant, fontSize: 12),
+                  ),
+                  Text(
+                    _formatCurrency(offer.totalCents),
+                    style: const TextStyle(
+                      fontWeight: FontWeight.bold,
+                      fontSize: 18,
+                      color: Color(0xFF00A84E),
+                    ),
+                  ),
+                ],
+              ),
+              ElevatedButton(
+                onPressed: onBuyPressed,
+                style: ElevatedButton.styleFrom(
+                  backgroundColor: const Color(0xFF00A84E),
+                  foregroundColor: Colors.white,
+                  elevation: 0,
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(8),
+                  ),
+                ),
+                child: const Text('Comprar'),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildStatusBadge(OfferStatus status) {
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
+      decoration: BoxDecoration(
+        color: _getStatusColor(status).withOpacity(0.1),
+        borderRadius: BorderRadius.circular(6),
+      ),
+      child: Text(
+        status.toDisplayString(),
+        style: TextStyle(
+          color: _getStatusColor(status),
+          fontWeight: FontWeight.bold,
+          fontSize: 12,
+        ),
+      ),
+    );
+  }
+
+  Widget _buildInfoItem(BuildContext context, IconData icon, String value, {String? label}) {
+    final theme = Theme.of(context);
+    return Row(
+      children: [
+        Icon(icon, size: 16, color: theme.colorScheme.onSurfaceVariant),
+        const SizedBox(width: 4),
+        Text(
+          value,
+          style: TextStyle(
+            color: theme.colorScheme.onSurface,
+            fontWeight: FontWeight.w500,
+          ),
+        ),
+        if (label != null) ...[
+          const SizedBox(width: 2),
+          Text(
+            label,
+            style: TextStyle(color: theme.colorScheme.onSurfaceVariant, fontSize: 12),
+          ),
+        ],
+      ],
+    );
+  }
+}
