--- frontend/lib/widgets/offers_view.dart
@@ -1,521 +1,0 @@
-// Autor: Allan Giovanni Matias Paes
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
-import 'package:frontend/models/offer.dart';
-import 'package:frontend/services/offer_service.dart';
-import 'package:frontend/pages/my_offers_page.dart';
-import 'package:frontend/pages/buy_offer_page.dart';
-import 'package:frontend/pages/create_offer_page.dart';
-import 'package:frontend/widgets/feedback_modal.dart';
-import 'package:intl/intl.dart';
-import 'package:shimmer/shimmer.dart';
-import 'package:flutter/services.dart';
-
-// Visão que exibe todas as ofertas abertas de tokens no mercado secundário.
-class OffersView extends StatefulWidget {
-  const OffersView({super.key});
-
-  @override
-  State<OffersView> createState() => _OffersViewState();
-}
-
-class _OffersViewState extends State<OffersView> {
-  final ScrollController _scrollController = ScrollController();
-  final List<OfferWithId> _offers = [];
-  bool _isLoading = false;
-  bool _hasMore = true;
-  String? _lastOfferId;
-  String _searchStartup = "";
-  double? _maxPrice;
-  final TextEditingController _priceController = TextEditingController();
-  final String? _currentUserId = FirebaseAuth.instance.currentUser?.uid;
-  
-  final NumberFormat _currencyFormat = NumberFormat.currency(
-    locale: 'pt_BR',
-    symbol: 'R\$',
-  );
-
-  @override
-  void initState() {
-    super.initState();
-    _loadMoreOffers();
-    _scrollController.addListener(_onScroll);
-  }
-
-  @override
-  void dispose() {
-    _scrollController.dispose();
-    _priceController.dispose();
-    super.dispose();
-  }
-
-  // Monitora o scroll para carregar mais itens quando o usuário chegar ao final da lista.
-  void _onScroll() {
-    if (_scrollController.position.pixels >=
-        _scrollController.position.maxScrollExtent - 200) {
-      _loadMoreOffers();
-    }
-  }
-
-  // Busca mais ofertas no servidor, gerenciando o estado de carregamento e paginação.
-  Future<void> _loadMoreOffers({bool refresh = false}) async {
-    if (_isLoading || (!_hasMore && !refresh)) return;
-
-    setState(() {
-      _isLoading = true;
-      if (refresh) {
-        _offers.clear();
-        _lastOfferId = null;
-        _hasMore = true;
-      }
-    });
-
-    final result = await OfferService.getOffers(
-      limit: 15,
-      startAfter: _lastOfferId,
-    );
-
-    if (mounted) {
-      if (result.success) {
-        final List<OfferWithId> newOffers = result.data!['offers'];
-        final String? lastId = result.data!['lastOfferId'];
-
-        setState(() {
-          _offers.addAll(newOffers);
-          _lastOfferId = lastId;
-          _isLoading = false;
-          if (newOffers.length < 15 || lastId == null) {
-            _hasMore = false;
-          }
-        });
-      } else {
-        setState(() => _isLoading = false);
-        FeedbackModal.show(
-          context: context,
-          title: 'Erro ao carregar',
-          message: result.message ?? 'Erro ao carregar ofertas',
-          type: FeedbackType.error,
-        );
-      }
-    }
-  }
-
-  // Filtra a lista de ofertas carregadas localmente com base na busca e preço máximo.
-  List<OfferWithId> get _filteredOffers {
-    return _offers.where((offer) {
-      // Não exibe ofertas criadas pelo próprio usuário logado.
-      if (_currentUserId != null && offer.seller.id == _currentUserId) {
-        return false;
-      }
-
-      final matchesStartup = offer.startupName
-          .toLowerCase()
-          .contains(_searchStartup.toLowerCase());
-      
-      // Converte o preço de centavos para reais para comparar com o filtro de preço máximo.
-      final matchesPrice = _maxPrice == null || 
-          (offer.tokenPriceCents / 100) <= _maxPrice!;
-          
-      return matchesStartup && matchesPrice;
-    }).toList();
-  }
-
-  // Auxiliar para formatar valores em centavos para a moeda Real (R$).
-  String _formatCurrency(int cents) {
-    return _currencyFormat.format(cents / 100);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Scaffold(
-      backgroundColor: theme.scaffoldBackgroundColor,
-      appBar: AppBar(
-        title: Text(
-          'Mercado',
-          style: TextStyle(fontWeight: FontWeight.bold, color: theme.colorScheme.onSurface),
-        ),
-        actions: [
-          IconButton(
-            icon: Icon(Icons.list_alt, color: theme.colorScheme.onSurface),
-            onPressed: () {
-              Navigator.of(context).push(
-                MaterialPageRoute(builder: (context) => const MyOffersView()),
-              );
-            },
-            tooltip: 'Minhas Ofertas',
-          ),
-        ],
-        backgroundColor: theme.scaffoldBackgroundColor,
-        elevation: 0,
-        centerTitle: true,
-      ),
-      body: Column(
-        children: [
-          _buildFilters(),
-          Expanded(
-            child: RefreshIndicator(
-              onRefresh: () => _loadMoreOffers(refresh: true),
-              child: _offers.isEmpty && _isLoading
-                  ? _buildSkeletonLoading()
-                  : _filteredOffers.isEmpty
-                      ? _buildEmptyState()
-                      : ListView.builder(
-                          controller: _scrollController,
-                          physics: const AlwaysScrollableScrollPhysics(),
-                          itemCount: _filteredOffers.length + (_hasMore ? 1 : 0),
-                          padding: const EdgeInsets.all(16.0),
-                          itemBuilder: (context, index) {
-                            if (index == _filteredOffers.length) {
-                              return const Center(
-                                child: Padding(
-                                  padding: EdgeInsets.all(16.0),
-                                  child: CircularProgressIndicator(),
-                                ),
-                              );
-                            }
-                            final offer = _filteredOffers[index];
-                            return _buildOfferCard(offer);
-                          },
-                        ),
-            ),
-          ),
-        ],
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () async {
-          final result = await Navigator.of(context).push<bool>(
-            MaterialPageRoute(builder: (context) => const CreateOfferPage()),
-          );
-          if (result == true) {
-            _loadMoreOffers(refresh: true);
-          }
-        },
-        backgroundColor: const Color(0xFF00A84E),
-        child: const Icon(Icons.add, color: Colors.white),
-      ),
-    );
-  }
-
-  // Constrói os campos de filtro de busca por startup e por preço máximo.
-  Widget _buildFilters() {
-    final theme = Theme.of(context);
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
-      child: Row(
-        children: [
-          Expanded(
-            flex: 2,
-            child: TextField(
-              onChanged: (value) {
-                setState(() {
-                  _searchStartup = value;
-                });
-              },
-              style: TextStyle(color: theme.colorScheme.onSurface),
-              decoration: InputDecoration(
-                hintText: 'Startup...',
-                hintStyle: TextStyle(color: theme.colorScheme.onSurfaceVariant.withOpacity(0.5)),
-                prefixIcon: const Icon(Icons.search, color: Color(0xFF00A84E)),
-                filled: true,
-                fillColor: theme.colorScheme.surfaceVariant.withOpacity(0.3),
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(12),
-                  borderSide: BorderSide.none,
-                ),
-                contentPadding: const EdgeInsets.symmetric(vertical: 0),
-              ),
-            ),
-          ),
-          const SizedBox(width: 8),
-          Expanded(
-            flex: 1,
-            child: TextField(
-              controller: _priceController,
-              keyboardType: TextInputType.number,
-              style: TextStyle(color: theme.colorScheme.onSurface),
-              inputFormatters: [
-                FilteringTextInputFormatter.allow(RegExp(r'^\d*[.,]?\d{0,2}')),
-              ],
-              onChanged: (value) {
-                setState(() {
-                  // Normaliza vírgula para ponto antes de converter para double.
-                  String normalizedValue = value.replaceAll(',', '.');
-                  _maxPrice = double.tryParse(normalizedValue);
-                });
-              },
-              decoration: InputDecoration(
-                hintText: 'Máx R\$',
-                hintStyle: TextStyle(color: theme.colorScheme.onSurfaceVariant.withOpacity(0.5)),
-                filled: true,
-                fillColor: theme.colorScheme.surfaceVariant.withOpacity(0.3),
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(12),
-                  borderSide: BorderSide.none,
-                ),
-                contentPadding: const EdgeInsets.symmetric(horizontal: 12, vertical: 0),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  // Constrói o card individual de cada oferta com detalhes de quantidade, preço e ação de compra.
-  Widget _buildOfferCard(OfferWithId offer) {
-    final theme = Theme.of(context);
-    return Container(
-      margin: const EdgeInsets.only(bottom: 16),
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: theme.colorScheme.surface,
-        borderRadius: BorderRadius.circular(16),
-        border: Border.all(color: theme.dividerColor.withOpacity(0.1)),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.04),
-            blurRadius: 10,
-            offset: const Offset(0, 4),
-          ),
-        ],
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      offer.startupName,
-                      style: TextStyle(
-                        fontWeight: FontWeight.bold,
-                        fontSize: 18,
-                        color: theme.colorScheme.onSurface,
-                      ),
-                    ),
-                    const SizedBox(height: 2),
-                    Text(
-                      'Vendedor: ${offer.seller.name}',
-                      style: TextStyle(
-                        fontSize: 12,
-                        color: theme.colorScheme.onSurfaceVariant,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              _buildStatusBadge(offer.status),
-            ],
-          ),
-          const SizedBox(height: 12),
-          Row(
-            children: [
-              _buildInfoItem(
-                Icons.token_outlined,
-                '${offer.qtdTokens} tokens',
-              ),
-              const SizedBox(width: 16),
-              _buildInfoItem(
-                Icons.monetization_on_outlined,
-                _formatCurrency(offer.tokenPriceCents),
-                label: 'cada',
-              ),
-            ],
-          ),
-          Divider(height: 24, color: theme.dividerColor.withOpacity(0.1)),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    'Total da Oferta',
-                    style: TextStyle(color: theme.colorScheme.onSurfaceVariant, fontSize: 12),
-                  ),
-                  Text(
-                    _formatCurrency(offer.totalCents),
-                    style: const TextStyle(
-                      fontWeight: FontWeight.bold,
-                      fontSize: 18,
-                      color: Color(0xFF00A84E),
-                    ),
-                  ),
-                ],
-              ),
-              ElevatedButton(
-                onPressed: () async {
-                  // Exibe indicador de carregamento enquanto verifica a expiração da oferta.
-                  showDialog(
-                    context: context,
-                    barrierDismissible: false,
-                    builder: (context) => const Center(child: CircularProgressIndicator(color: Color(0xFF00A84E))),
-                  );
-
-                  final resultCheck = await OfferService.isOfferExpired(offerId: offer.id);
-
-                  if (context.mounted) {
-                    Navigator.of(context).pop(); // Fecha o indicador de carregamento.
-                  }
-
-                  if (resultCheck.success) {
-                    if (resultCheck.data == true) {
-                      if (context.mounted) {
-                        FeedbackModal.show(
-                          context: context,
-                          title: 'Oferta Expirada',
-                          message: 'Essa oferta acabou de expirar! Que tal conferir outras oportunidades no catálogo?',
-                          type: FeedbackType.info,
-                          onConfirm: () => _loadMoreOffers(refresh: true),
-                        );
-                      }
-                    } else {
-                      if (context.mounted) {
-                        final result = await Navigator.of(context).push<bool>(
-                          MaterialPageRoute(
-                            builder: (context) => BuyOfferPage(offer: offer),
-                          ),
-                        );
-                        if (result == true) {
-                          _loadMoreOffers(refresh: true);
-                        }
-                      }
-                    }
-                  } else {
-                    if (context.mounted) {
-                      FeedbackModal.show(
-                        context: context,
-                        title: 'Erro ao Verificar',
-                        message: resultCheck.message ?? 'Não foi possível verificar o status da oferta',
-                        type: FeedbackType.error,
-                      );
-                    }
-                  }
-                },
-                style: ElevatedButton.styleFrom(
-                  backgroundColor: const Color(0xFF00A84E),
-                  foregroundColor: Colors.white,
-                  elevation: 0,
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                ),
-                child: const Text('Comprar'),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  // Helper para construir itens de informação com ícone e texto.
-  Widget _buildInfoItem(IconData icon, String value, {String? label}) {
-    final theme = Theme.of(context);
-    return Row(
-      children: [
-        Icon(icon, size: 16, color: theme.colorScheme.onSurfaceVariant),
-        const SizedBox(width: 4),
-        Text(
-          value,
-          style: TextStyle(
-            color: theme.colorScheme.onSurface,
-            fontWeight: FontWeight.w500,
-          ),
-        ),
-        if (label != null) ...[
-          const SizedBox(width: 2),
-          Text(
-            label,
-            style: TextStyle(color: theme.colorScheme.onSurfaceVariant, fontSize: 12),
-          ),
-        ],
-      ],
-    );
-  }
-
-  // Badge colorido que indica o status atual da oferta.
-  Widget _buildStatusBadge(OfferStatus status) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-      decoration: BoxDecoration(
-        color: _getStatusColor(status).withOpacity(0.1),
-        borderRadius: BorderRadius.circular(6),
-      ),
-      child: Text(
-        status.toDisplayString(),
-        style: TextStyle(
-          color: _getStatusColor(status),
-          fontWeight: FontWeight.bold,
-          fontSize: 12,
-        ),
-      ),
-    );
-  }
-
-  // Retorna a cor correspondente a cada status de oferta.
-  Color _getStatusColor(OfferStatus status) {
-    switch (status) {
-      case OfferStatus.open:
-        return Colors.green;
-      case OfferStatus.accepted:
-        return Colors.blue;
-      case OfferStatus.cancelled:
-        return Colors.red;
-      case OfferStatus.expired:
-        return Colors.orange;
-    }
-  }
-
-  Widget _buildSkeletonLoading() {
-    final isDark = Theme.of(context).brightness == Brightness.dark;
-    return ListView.builder(
-      itemCount: 5,
-      padding: const EdgeInsets.all(16.0),
-      itemBuilder: (context, index) {
-        return Shimmer.fromColors(
-          baseColor: isDark ? Colors.grey[800]! : Colors.grey[300]!,
-          highlightColor: isDark ? Colors.grey[700]! : Colors.grey[100]!,
-          child: Container(
-            height: 150,
-            margin: const EdgeInsets.only(bottom: 16),
-            decoration: BoxDecoration(
-              color: Colors.white,
-              borderRadius: BorderRadius.circular(16),
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  Widget _buildEmptyState() {
-    final theme = Theme.of(context);
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        return SingleChildScrollView(
-          physics: const AlwaysScrollableScrollPhysics(),
-          child: SizedBox(
-            height: constraints.maxHeight,
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Icon(Icons.local_offer_outlined, size: 64, color: theme.colorScheme.onSurfaceVariant.withOpacity(0.3)),
-                const SizedBox(height: 16),
-                Text(
-                  'Nenhuma oferta encontrada',
-                  style: TextStyle(color: theme.colorScheme.onSurfaceVariant, fontSize: 16),
-                ),
-              ],
-            ),
-          ),
-        );
-      }
-    );
-  }
-}
